Implement ftp retry for update_ted management command #232

+++ app/parsers/ted.py
@@ -53,7 +53,6 @@ class TEDWorker:
         while last_date <= today:
             self.download_archive(ftp, last_date, archives)
 
-
             last_date += timedelta(1)
 
             if last_year != last_date.strftime("%Y"):
@@ -95,11 +94,17 @@ class TEDWorker:
             if not os.path.exists(self.path):
                 os.makedirs(self.path)
             file_path = os.path.join(self.path, archive_name)
-            with open(file_path, "wb") as f:
-                ftp.retrbinary(
-                    "RETR %s" % archive_name, lambda data: f.write(data)
-                )
-            self.archives.append(file_path)
+            while True:
+                if not os.path.exists(file_path):
+                    with open(file_path, "wb") as f:
+                        ftp.retrbinary(
+                            "RETR %s" % archive_name, lambda data: f.write(data)
+                        )
+                    self.archives.append(file_path)
+                    break
+                logging.warning(f"File already downloaded, waiting 30 seconds: {file_path}")
+                time.sleep(30)
+                logging.warning("Retrying")
 
     def parse_notices(self, tenders=[], set_notified=False):
         changed_tenders = []
@@ -147,20 +152,15 @@ class TEDWorker:
     def ftp_login():
         while True:
             try:
-                import pdb
-                pdb.set_trace()
                 ftp = FTP(settings.TED_FTP_URL)
                 ftp.login(user=settings.TED_FTP_USER, passwd=settings.TED_FTP_PASSWORD)
                 logging.warning("Logged into FTP.")
-                break
+                return ftp
             except error_perm as e:
-                logging.warning(f"Waiting 30 seconds: {e}")
+                logging.warning(f"Cannot login to FTP, waiting 30 seconds: {e}")
                 time.sleep(30)
                 logging.warning("Retrying")
                 continue
-
-
-        return ftp
 
     @staticmethod
     def get_archive_name(last_date, archives):
